feat: add field for minimum CLI version
This is preparing for a future feature. A Cloud Assembly producer
can't yet know what CLI version would be required to read it,
but in the future once we tie the Cloud Assembly package version
to the CLI package version, we will be able to do that.

If we build the "check and report" feature in right now, at the time we
start producing manifests with this field set, more CLI versions will be
out that will be taking advantage of it, and reporting useful error
messages.