chore: store schema version inside package (#96)
(Take 2 of #62).

Currently, the schema version espoused by this package is the package version itself (i.e., the version of its own `package.json`). Move this version to a resource file that maintained inside the package and is bumped explicitly.

In a projen (mono)repo, all packages have version 0.0.0. At development time, the Cloud Assembly Schema has to know its own version, for various versioning checks. It can’t get it from the package version (which is 0.0.0, so it needs to get it elsewhere, which will be schema.version.json).

Closes #62.